services/data-analytics/requirements.txt
paho-mqtt==1.6.1
numpy==1.24.3
scikit-learn==1.3.0
python-dotenv==1.0.0
